Put in place a way to prevent CodeParser from being run from source.
Because CodeParser has a native library to be built, it is a problem if you just call PackPaclet[codeParserSourceDir].

I have changed the PacletInfo.m file in source to have a UnbuiltPaclet[] wrapper instead of a Paclet[] wrapper. This then gets replaced during the build process.

This prevents the source repo from being treated as a paclet.